requirements.txt
python-docx>=0.8.11
pyinstaller>=5.13.0
Pillow>=10.0.0
